% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Countries-package.R
\docType{package}
\name{Countries-package}
\alias{Countries}
\alias{Countries-package}
\title{Countries: Find the countrries I have lived in.}
\description{
I have lived in 10 countries.
}
\author{
\strong{Maintainer}: Patrick Kelly \email{phototala@gmail.com}

}
\keyword{internal}
