% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\docType{package}
\name{countries}
\alias{countries}
\alias{countries-package}
\title{A tool for generating lists of country names based on various definitions
and organising them into different administrative, socio-economic, and/or
political groupings.}
\description{
A tool for generating lists of country names based on various definitions and organising them into different administrative, socio-economic, and/or political groupings.
}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

}
\keyword{internal}
