% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{country}
\alias{country}
\title{Find a country}
\usage{
country(name)
}
\arguments{
\item{name}{A country name provided as a character string.}
}
\description{
A country I have lived in.
}
\examples{
country("Sierra Leone")
}
