% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_name.R
\name{country_name}
\alias{country_name}
\title{Convert and translate country names}
\usage{
country_name(
  x,
  to = "ISO3",
  fuzzy_match = TRUE,
  verbose = TRUE,
  simplify = TRUE,
  custom_table = NULL
)
}
\arguments{
\item{x}{A vector of country names}

\item{to}{A string containing the desired naming conventions to which \code{x} should be converted to. Possible values are: "all", "ISO-3", "ISO-2", ... Default is "ISO3".}

\item{fuzzy_match}{Logical value indicating whether fuzzy matching of country names should be allowed (\code{TRUE}), or only exact matches are allowed (\code{FALSE}). Default is \code{TRUE}.}

\item{verbose}{Logical value indicating whether the function should print to the console a report on the matching process. Default is \code{FALSE}.}

\item{simplify}{Logical value. If set to \code{TRUE} the function will return the match table as a \code{data.frame} object. If set to \code{FALSE}, the function will return a list object containing the match table and additional details on the country matching process. Default is \code{TRUE}.}

\item{custom_table}{Custom conversion table to be used. This needs to be a data.frame object. Default is \code{NULL}.}
}
\value{
Returns a vector of converted country names.
}
\description{
This function recognises and converts country names to different nomenclatures and languages.
It supports fuzzy matching for greater flexibility.
}
\examples{
country_name(x=c("US","Italia","France","United States"), to= "ISO3")
}
