% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_reference_list}
\alias{country_reference_list}
\title{Conversion table}
\format{
A data frame with columns corresponding to different country naming conventions.
\describe{
  \item{ISO3}{ISO-3 country codes}
}
}
\usage{
country_reference_list
}
\description{
A table containing country names in different naming conventions
}
\keyword{datasets}
