% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_countrycol.R
\name{find_countrycol}
\alias{find_countrycol}
\title{Finds columns containing country names}
\usage{
find_countrycol(x, return_index = FALSE)
}
\arguments{
\item{x}{A data frame object}

\item{return_index}{A logical value indicating whether the function should return the index of country columns instead of the column names. Default is \code{FALSE}, column names are returned.}
}
\value{
Returns a vector of country names (\code{return_index=FALSE}) or column indices (\code{return_index=TRUE}) of columns containing country names.
}
\description{
This function takes a data frame as argument and returns the column name (or index) of all columns containing country names.
It can be used to automate the search of country columns in data frames.
For the purpose of this function, a country is any of the 249 territories designated in the ISO standard \code{3166}.
}
\examples{
find_countrycol(x=data.frame(a=c("Brésil","Tonga","FRA"), b=c(1,2,3)))
}
\seealso{
\link[countries]{is_country} \link[countries]{country_name} \link[countries]{find_timecol}
}
