% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has.invalid.multibyte.string.R
\name{has.invalid.multibyte.string}
\alias{has.invalid.multibyte.string}
\title{Internal function - Determine if "invalid multibyte string" error will be triggered due to wrong encoding.}
\usage{
has.invalid.multibyte.string(x, return.elements = FALSE)
}
\arguments{
\item{x}{String vector to be tested}

\item{return.elements}{Logical value. If `TRUE`, then the function returns a logical value for each element of x, otherwise a single logical is returned for the entire vector.}
}
\description{
Internal function - Determine if "invalid multibyte string" error will be triggered due to wrong encoding.
}
