% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_date.R
\name{is.date}
\alias{is.date}
\title{internal function - Tests whether the value is convertible to a date format}
\usage{
is.date(x, formats)
}
\arguments{
\item{x}{A vector of length 1 to be tested}

\item{formats}{Formats to check for expressed in standard notation.}
}
\value{
Returns a logical value indicating whether the value can be converted to any of the date formats provided. This function is the basis for \code{countries::is_date()}.
}
\description{
Internal function behind \code{is_date()}.
}
\examples{
is.date("2020-01-01")
is.yearcol("a")
}
\seealso{
\link[countries]{is_date} \link[countries]{find_timecol}
}
