% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.yearcol.R
\name{is.yearcol}
\alias{is.yearcol}
\title{Internal function - Tests whether a vector of data could be containing year information}
\usage{
is.yearcol(x, limit = c(1800, 2200), regularity = FALSE, allow_NA = TRUE)
}
\arguments{
\item{x}{A vector to be tested}

\item{limit}{A minimum and maximum year value \code{c(1800,2200)}}

\item{regularity}{Require that no gap is present in the year values (e.g. 2005, 2000, 2010, 2020 would return FALSE because 2015 is missing). Default is \code{FALSE} (regularity is not required).}

\item{allow_NA}{If set to true the date vector is allowed to contain \code{NA} values. If set to \code{FALSE} the function returns a FALSE if any value is NA. Default is \code{TRUE}.}
}
\value{
Returns a logical vector indicating whether the vector could be containing year information
}
\description{
Internal function - Tests whether a vector of data could be containing year information
}
\examples{
countries:::is.yearcol(1990:2020)
countries:::is.yearcol(c("a",2000))
}
\seealso{
\link[countries]{is_country} \link[countries]{find_countrycol}
}
