% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{mode}
\alias{mode}
\title{Get the mode of a vector}
\usage{
mode(x, first_only = FALSE)
}
\arguments{
\item{x}{A vector}

\item{first_only}{Logical value indicating whether only the first mode should be returned in \code{x} has multiple modes (i.e. there are multiple values with the highest number of observations). Default is FALSE.}
}
\value{
Returns the mode of the vector \code{x}
}
\description{
This function returns the mode of a distribution. For any given vector of values, it returns the value that appears most frequently.
The function works with strings, numerical and mixed inputs. \code{NA} values are treated as a distinct value.
}
\examples{
countries::mode(c("a","a",2,3))
countries::mode(c(1,1,2,3,NA,2))
}
