% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_palettes_and_themes.R
\name{palettes_countries}
\alias{palettes_countries}
\title{Discrete colour palettes}
\usage{
palettes_countries(n, theme = 1, reverse = FALSE)
}
\arguments{
\item{n}{Number of desired colours}

\item{theme}{A numeric value or name identifying the theme's colours. Can be a number between 1 and 11, or one of the theme's names: \code{c("Default", "Greyscale", "Candy", "RedBlue", "Dark", "Reds", "Blues", "Greens", "Viridis", "Cividis", "Distinct", "Distinct2", "Paired")}.}

\item{reverse}{Logical value indicating whether to reverse the order of the palette's colours Default is FALSE.}
}
\value{
Returns n colours from the requested theme
}
\description{
This function provides access to the discrete colours palettes used in this packages' 11 themes.
}
\examples{
palettes_countries(5, theme = 1)
}
\seealso{
\link[countries]{quick_map}
}
