% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_palettes_and_themes.R
\name{pick_colours}
\alias{pick_colours}
\title{Internal function - return n spaced indices for picking colours from a palette}
\usage{
pick_colours(n, n_colours_palette)
}
\arguments{
\item{n}{Desired Number of colours desired}

\item{n_colours_palette}{Total number of colours in the palette}
}
\value{
Returns n indices that can be used to select colours from a palette of n_colours_palette length
}
\description{
Internal function - return n spaced indices for picking colours from a palette
}
\examples{
countries:::pick_colours(5, 10)
}
\seealso{
\link[countries]{palettes_countries}
}
