% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_palettes_and_themes.R
\name{themes_countries}
\alias{themes_countries}
\title{Internal function - ggplot themes for quick_map}
\usage{
themes_countries(theme = 1)
}
\arguments{
\item{theme}{A numeric value or name identifying the theme's colours. Can be a number between 1 and 11, or one of the theme's names: \code{c("Default", "Greyscale", "Candy", "RedBlue", "Dark", "Reds", "Blues", "Greens", "Viridis", "Cividis", "Distinct")}.}
}
\value{
Returns n indices that can be used to select colours from a palette of n_colours_palette length
}
\description{
Internal function - ggplot themes for quick_map
}
\examples{
countries:::themes_countries(theme = 1)
}
\seealso{
\link[countries]{quick_map}
}
