% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mode.R
\name{get_mode}
\alias{get_mode}
\title{Get the mode of a vector}
\usage{
get_mode(x)
}
\arguments{
\item{x}{A vector}
}
\value{
Returns the mode of the vector \code{x}
}
\description{
This function returns the mode of a distribution. For any given vector of values, it returns the value that appears most frequently.
The function works with strings, numerical and mixed inputs. \code{NA} values are treated as distinct value.
In case there are two or more values that appear most frequently in the vector, only the first is returned.
}
\examples{
get_mode(c("a","a",2,3))
get_mode(c(1,1,2,3,NA))
}
