% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.by.order.R
\name{parse.by.order}
\alias{parse.by.order}
\title{Internal function - Parses input merging order for function auto_join}
\usage{
parse.by.order(data, by, col_names)
}
\arguments{
\item{data}{list of tables provided to auto_join function}

\item{by}{by order provided by user to auto_join function}

\item{col_names}{list of column names in \code{data} tables}
}
\value{
Returns a cleaned \code{by} order and a vector indicating the type of data in each element of the order.
}
\description{
Internal function - Parses input merging order for function auto_join
}
